fuzz

+++ src/Wallet.sol
@@ -9,13 +9,12 @@ contract Wallet {
 
     receive() external payable {}
 
-    function withdraw(uint _amount) external {
+    function withdraw(uint256 _amount) external {
         require(msg.sender == owner, "caller is not owner");
         payable(msg.sender).transfer(_amount);
     }
 
-    function getBalance() external view returns (uint) {
+    function getBalance() external view returns (uint256) {
         return address(this).balance;
     }
 }
-
