+++ test/Error.t.sol
@@ -32,10 +32,10 @@ contract ErrorTest is Test {
 
     // Add label to assertions
     function testErrorLabel() public {
-        assertEq(uint(1), uint(1), "test 1");
-        assertEq(uint(1), uint(1), "test 2");
-        assertEq(uint(1), uint(1), "test 3");
-        assertEq(uint(1), uint(1), "test 4");
-        assertEq(uint(1), uint(1), "test 5");
+        assertEq(uint256(1), uint256(1), "test 1");
+        assertEq(uint256(1), uint256(1), "test 2");
+        assertEq(uint256(1), uint256(1), "test 3");
+        assertEq(uint256(1), uint256(1), "test 4");
+        assertEq(uint256(1), uint256(1), "test 5");
     }
 }
